+++ lib/openstudio/lib/openstudio_aws_wrapper.rb
@@ -75,30 +75,29 @@ class OpenStudioAwsWrapper
     @aws = Aws::EC2::Client.new(options[:credentials])
   end
 
+  # TODO: Security overhaul: lock down the groups to only specific ports
   def create_or_retrieve_default_security_group
     tmp_name = 'openstudio-server-sg-v1'
     group = @aws.describe_security_groups(filters: [{ name: 'group-name', values: [tmp_name] }])
     logger.info "Length of the security group is: #{group.data.security_groups.length}"
     if group.data.security_groups.length == 0
-      logger.info 'server group not found --- will create a new one'
+      logger.info 'security group not found --- will create a new one'
       @aws.create_security_group(group_name: tmp_name, description: "group dynamically created by #{__FILE__}")
       @aws.authorize_security_group_ingress(
           group_name: tmp_name,
           ip_permissions: [
-            { ip_protocol: 'tcp', from_port: 1, to_port: 65_535, ip_ranges: [cidr_ip: '0.0.0.0/0'] }
-          ]
-      )
-      @aws.authorize_security_group_ingress(
-          group_name: tmp_name,
-          ip_permissions: [
-            { ip_protocol: 'icmp', from_port: -1, to_port: -1, ip_ranges: [cidr_ip: '0.0.0.0/0']
-            }
+            { ip_protocol: 'tcp', from_port: 1, to_port: 65535, ip_ranges: [cidr_ip: '0.0.0.0/0'] },
+            { ip_protocol: 'icmp', from_port: -1, to_port: -1, ip_ranges: [cidr_ip: '0.0.0.0/0'] }
+             # R used 11000 - 12000
           ]
       )
 
       # reload group information
       group = @aws.describe_security_groups(filters: [{ name: 'group-name', values: [tmp_name] }])
+    else
+      logger.info "Found existing security group"
     end
+
     @security_groups = [group.data.security_groups.first.group_id]
     logger.info("server_group #{group.data.security_groups.first.group_name}:#{group.data.security_groups.first.group_id}")
   end
@@ -304,6 +303,7 @@ class OpenStudioAwsWrapper
 
     # replace the server_script.sh.template with the keys to add
     user_data = File.read(File.expand_path(File.dirname(__FILE__)) + '/server_script.sh.template')
+    user_data.gsub!(/SERVER_HOSTNAME/, 'openstudio.server')
     user_data.gsub!(/WORKER_PRIVATE_KEY_TEMPLATE/, worker_keys.private_key.gsub("\n","\\n"))
     user_data.gsub!(/WORKER_PUBLIC_KEY_TEMPLATE/, worker_keys.ssh_public_key)
 
@@ -329,8 +329,7 @@ class OpenStudioAwsWrapper
 
     user_data = File.read(File.expand_path(File.dirname(__FILE__)) + '/worker_script.sh.template')
     user_data.gsub!(/SERVER_IP/, @server.data.private_ip_address)
-    user_data.gsub!(/SERVER_HOSTNAME/, 'master')
-    user_data.gsub!(/SERVER_ALIAS/, '')
+    user_data.gsub!(/SERVER_HOSTNAME/, 'openstudio.server')
     user_data.gsub!(/WORKER_PUBLIC_KEY_TEMPLATE/, worker_keys.ssh_public_key)
     logger.info("worker user_data #{user_data.inspect}")
 
@@ -355,7 +354,9 @@ class OpenStudioAwsWrapper
   end
 
   # blocking method that waits for servers and workers to be fully configured (i.e. execution of user_data has
-  # occured on all nodes)
+  # occured on all nodes). Ideally none of these methods would ever need to exist.
+  #
+  # @return [Boolean] Will return true unless an exception is raised
   def configure_server_and_workers
     logger.info('waiting for server user_data to complete')
     @server.wait_command('[ -e /home/ubuntu/user_data_done ] && echo "true"')
@@ -372,12 +373,6 @@ class OpenStudioAwsWrapper
     logger.info("ips #{ips}")
     @server.shell_command('chmod 664 /home/ubuntu/ip_addresses')
 
-    # TODO: Grab a bunch of AWS specific information and save into a file in home directory
-
-
-    # @server.shell_command('~/setup-ssh-keys.sh')
-    # @server.shell_command('~/setup-ssh-worker-nodes.sh ip_addresses')
-
     mongoid = File.read(File.expand_path(File.dirname(__FILE__)) + '/mongoid.yml.template')
     mongoid.gsub!(/SERVER_IP/, @server.data.private_ip_address)
     file = Tempfile.new('mongoid.yml')
@@ -390,8 +385,6 @@ class OpenStudioAwsWrapper
     @server.shell_command('chmod 664 /mnt/openstudio/rails-models/mongoid.yml')
     @workers.each { |worker| worker.shell_command('chmod 664 /mnt/openstudio/rails-models/mongoid.yml') }
 
-    # I'm removing this as of 10/6/14. This should have been resolved by now.
-    # sleep 60 # wait 60 more seconds for everything -- this is cheesy
     true
   end
 
@@ -545,8 +538,13 @@ class OpenStudioAwsWrapper
         a[:tested] = false
       end
 
-      # TODO: in 1.6.0 just stop putting in the cc2workers.
-      if ami[:tags_hash][:openstudio_version].to_version >= '1.5.0'
+      if ami[:tags_hash][:openstudio_version].to_version >= '1.6.0'
+        if ami[:name] =~ /Server/
+          a[:amis][:server] = ami[:image_id]
+        elsif ami[:name] =~ /Worker/
+          a[:amis][:worker] = ami[:image_id]
+        end
+      elsif ami[:tags_hash][:openstudio_version].to_version >= '1.5.0'
         if ami[:name] =~ /Server/
           a[:amis][:server] = ami[:image_id]
         elsif ami[:name] =~ /Worker/